YuchenZhou_jiaqi_Pipeline/preprocessing/requirements.txt
# Python dependencies for MIMIC-IV temporal preprocessing pipeline

# Core data processing
numpy>=1.21.0
pandas>=1.3.0
pyarrow>=5.0.0  # For Parquet support

# Configuration
pyyaml>=5.4.0

# Progress bars
tqdm>=4.62.0

# Data validation
scikit-learn>=0.24.0

# Utilities
python-dateutil>=2.8.0
pytz>=2021.1

# Optional: For visualization/debugging
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: For deep learning model integration
torch>=1.9.0  # PyTorch
